WIP: wait-update-current-buffer in Document_activate
WIP: Still trying to get clipboard-cut to actually remove the text that's
supposed to be getting cut from the document.

I think I sort of "fixed" the crash when tm_delete of a tree-pointer happens in
C++ by using a guardian, but then never calling them out in the
after-gc-hook. So there's a TODO for that now.

This is mostly usable, until you want to cut and paste regions with zcite
fields in them...