Fix drd-props, better colorized debug printouts,
I've tested an article with tm-zotero, with multiple sections and the
bibliography in it's own section. I tested setting part-mode to several, then
back to all. It worked fine. Typing can be slow near or in front of a zcite
with logging turned on. It speeds up with logging off.

TODO Testing cut and paste has revealed a bug regarding an undefined reference,
     and one of the links in the bibliography is always a pink question-mark
     right now, after cutting then pasting a zcite citation cluster.