Improvements to tm-zotero.ts, tm-zotero.scm
Commented off code for the tp-guardian and now call tree-pointer-detach myself
inside clear-tree-pointer. The guardian thing did not work as expected; I think
that the hook function just keeps running or never returns; Instead of fixing
that, just try it this way once, eliminating the use of the guardian object.