WIP: Can cut zsubCite's out of zcite's, and paste them back in
The subCite's can be cut (not yet copied) from out of a disactivated zcite, and
then pasted into the main text of the document as a full zcite. Next I will be
making it possible to merge zcite's by pasting a zcite into the middle of a
disactivated zcite.

When those are done, I'll implement clipboard-copy for zsubCite's.